refactor: restore use case layer boundaries
